% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_slopes.R
\name{get_slopes}
\alias{get_slopes}
\title{Extract response-contingency slopes}
\usage{
get_slopes(data = NULL, r_id = "r_id", t_id = "t_id",
  response = "response", contingency = "contingency", compress = F)
}
\arguments{
\item{data}{data to use when fitting mixed model}

\item{r_id}{column of rater idenitfiers}

\item{t_id}{column of target identifiers}

\item{response}{column of responses (e.g., ratings, scores)}

\item{contingency}{column containing values that ratings are contingent on}

\item{compress}{column to compress data by ('rater' or 'target'). Default is no compression.}
}
\value{
New data frame containing individual slopes and intercepts representing
the relationship between ratings and contingencies for each rater
}
\description{
Extract response-contingency slopes
}
\examples{
## get_slopes(sloper_exdat, r_id = "r_id", t_id = "t_id",
##           response = "strength_rating", contingency = "target_strength_ave", compress = "rater")
## get_slopes(sloper_exdat, "strength_rating","target_strength_ave")
}
