% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_slopes.R
\name{plot_slopes}
\alias{plot_slopes}
\title{Visualize response-contingency slopes}
\usage{
plot_slopes(data = NULL, r_id = "r_id", response = "response",
  contingency = "contingency", xname = NULL, yname = NULL,
  groupfactor = NULL, gflevels = NULL, linear = T)
}
\arguments{
\item{data}{data to use for plotting}

\item{r_id}{column of rater idenitfiers}

\item{response}{column of responses (y-axis)}

\item{contingency}{column containing values that ratings are contingent on (x-axis)}

\item{xname}{custom name for x-axis}

\item{yname}{custom name for y-axis}

\item{groupfactor}{factor/grouping variable (color and linetype)}

\item{gflevels}{vector of custom labels for factor levels}

\item{linear}{logical argument indicating whether to plot linear function or nonlinear}
}
\value{
plot of the individual slopes and overall slope
}
\description{
Visualize response-contingency slopes
}
\examples{
# plot_slopes(sloper_exdat, response = "strength_rating", contingency = "t_measured_strength",
#            yname = "Stregth Rating", xname = "Target Strength",
#            groupfactor = "ratersex", linear = F, gflevels = c("Men", "Women"))
# plot_slopes(sloper_exdat, response = "strength_rating", contingency = "t_measured_strength")
}
