% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{slope}
\alias{slope}
\title{Find the slope of a line}
\usage{
slope(x1, y1, x2, y2)
}
\arguments{
\item{x1}{a number from a point (x1, y1)}

\item{y1}{a number from a point (x1,y1)}

\item{x2}{a number from a point (x2, y2)}

\item{y2}{a number from a point (x2, y2)}
}
\value{
The slope of a line between two given points (\code{x1}, \code{y1}) and (\code{x2}, \code{y2})
}
\description{
Find the slope of a line
}
\examples{
slope(1,2,4,5)
slope(63, 54, 34, 67)
}
