% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sloper_exdat.R
\docType{data}
\name{sloper_exdat}
\alias{sloper_exdat}
\title{Strength ratings of 66 men and rater characteristics}
\format{A data frame with 12012 rows and 14 variables
\describe{
\item{t_id}{Target ID}
\item{r_id}{Rater ID}
\item{ratersex}{Sex of the rater, 0 = male; 1 = female}
\item{rater_formidability}{Self-perceived formidability of the rater}
\item{rater_grip_strength}{Grip strength of the rater}
\item{rater_height}{Height in feet of the rater}
\item{rater_weight}{Weight in lbs of the rater}
\item{rater_age}{Age of the rater}
\item{strength_rating}{Ratings of target strength}
\item{target_strength_ave}{Average of ratings for each target}
\item{t_height}{Height in inches for each target}
\item{t_measured_strength}{Z-scored chest/grip strength composite for each target}
\item{t_age}{Age in years for each target}
\item{t_weight}{Weight in lbs for each target}

}}
\usage{
sloper_exdat
}
\description{
Raters (N = 187) rated the strength of 66 target males and provided answers to a number of individual-differences questions
}
\keyword{datasets}
